Posted the first_article_portfolio_teenage_models.md
Added tags in the header, but I have no idea, if I did it correctly. Can
not see them anywhere. Certainly can not add tags enclosed in single or
double quotes. Comma separated seems to work with white space even?!!!

The article needs to be put in a new category, that I could not create.

Added The images, that are shown in the body and added a caption for
each one.

Also added a gallery image, that is displayed on the projects page.

All images were added to `assets/img/` directory.