Ported data_prep_1-4.md from old site
Status: Renders on my local server, with expected problems.

WARNING: I have not checked the content or the rendered output in
detail!!! Still needs to be done.

ERROR: code blocks are not okay (no rendering of the code block background
among other problems.)

WARNING: Header should be fine, not sure what the 'tags: ...' line will do. Can
not see the tags anywhere.

TODO: There are some old hrefs in the articles, that need to be updated, once
the site is live

Minimal math changes were needed to make it work.