Changes to _posts/math behind stepping
ERROR: Removed the category parameter from the header, since it is not defined
for blog posts and tried to change tags, that have white space in them
to have correctly display on the website, but that did not work. Even
with a '-' it would not work.
I could not change the display_tag to have a '-' instead of whitespace.
It always has a white space no matter what.